+++ Models/Rack.cpp
@@ -53,7 +53,7 @@ void Rack::SetRackTiles(vector<Tile> Tiles){
 }
 
 void Rack::RemoveTile(Tile tile){
-    for (int i=0; i < tiles_.size(); i++){
+    for (int i=0; i < (int) tiles_.size(); i++){
         if (tiles_[i].GetLetter() == tile.GetLetter()){
             tiles_.erase(tiles_.begin()+i);
             break;
@@ -62,7 +62,7 @@ void Rack::RemoveTile(Tile tile){
 }
 
 void Rack::RemoveTile(char letter){
-    for (int i=0; i < tiles_.size(); i++){
+    for (int i=0; i < (int) tiles_.size(); i++){
         if (tiles_[i].GetLetter() == letter){
             tiles_.erase(tiles_.begin()+i);
             break;
@@ -71,7 +71,7 @@ void Rack::RemoveTile(char letter){
 }
 
 Tile Rack::RemoveAndReturnTile(char letter){
-    for (int i=0; i < tiles_.size(); i++){
+    for (int i=0; i < (int) tiles_.size(); i++){
         if (tiles_[i].GetLetter() == letter){
             Tile t= tiles_[i];
             tiles_.erase(tiles_.begin()+i);
@@ -82,7 +82,7 @@ Tile Rack::RemoveAndReturnTile(char letter){
 
 string Rack::RackToString(){
     string RackString ="";
-    for (int i=0; i< tiles_.size(); i++){
+    for (int i=0; i< (int) tiles_.size(); i++){
         RackString += tiles_[i].GetLetter();
     }
     return RackString;
